Split worker file into worker and processor.
